moved map to widgets
